+++ R/se.gllvm.R
@@ -148,14 +148,14 @@ se.gllvm <- function(object, ...){
       }
       
       if(method=="LA" || (num.lv==0 && (is.null(object$params$row.params.random) && is.null(object$randomX)) && object$col.eff$col.eff!="random")){
-        covM <- try(MASS::ginv(sdr[incl,incl]))
-        if(inherits(covM, "try-error")) { stop("Standard errors for parameters could not be calculated, due to singular fit.\n") }
-        se <- try(sqrt(diag(abs(covM))))
+        cov.mat.mod <- try(MASS::ginv(sdr[incl,incl]))
+        if(inherits(cov.mat.mod, "try-error")) { stop("Standard errors for parameters could not be calculated, due to singular fit.\n") }
+        se <- try(sqrt(diag(abs(cov.mat.mod))))
         names(se) = names(object$TMBfn$par[incl])
         
         trpred<-try({
           if(num.lv > 0 || !is.null(object$params$row.params.random) || !is.null(object$randomX) || object$col.eff$col.eff == "random") {
-            sd.random <- sdrandom(objrFinal, covM, incl)
+            sd.random <- sdrandom(objrFinal, cov.mat.mod, incl)
             prediction.errors <- list()
             
             if(!is.null(object$params$row.params.random)){
@@ -176,7 +176,7 @@ se.gllvm <- function(object, ...){
         }, silent=TRUE)
         if(inherits(trpred, "try-error")) { cat("Prediction errors for random effects could not be calculated.\n") }
         
-        out$Hess <- list(Hess.full=sdr, incl=incl, cov.mat.mod=covM)
+        out$Hess <- list(Hess.full=sdr, incl=incl, cov.mat.mod=cov.mat.mod)
       } else {
         sds <- sqrt(abs(diag(sdr)))
         if(any(sds<1e-12))sds[sds<1e-12]<-1
@@ -375,7 +375,7 @@ se.gllvm <- function(object, ...){
             if(attr(trm, "intercept"))LHS <- c("(Intercept)", LHS)
             RHS <- form[[3]]
             
-            names(sigma)[iter:(iter+trmsize[1,re]-1)] <- paste0(LHS, "|", RHS)
+            names(sigma)[iter:(iter+trmsize[1,re]-1)] <- paste0(LHS, "|", deparse(RHS))
             
             iter <- iter + trmsize[1,re]
           }
@@ -445,7 +445,7 @@ se.gllvm <- function(object, ...){
           se.zetanew <- matrix(NA,nrow=p,ncol=K)
           o_ind <- c(1:ncol(object$y))[family%in%c("ordinal", "orderedBeta")]
           zetas <- object$TMBfn$par[names(object$TMBfn$par)=="zeta"]
-          zeta.cov <- cov.mat.mod[names(object$TMBfn$par)[incl]=="zeta",names(object$TMBfn$par)[incl]=="zeta"]
+          zeta.cov <- cov.mat.mod[names(object$TMBfn$par)[incl]=="zeta",names(object$TMBfn$par)[incl]=="zeta",drop=FALSE]
           idx<-0
           for(j in o_ind){
             
@@ -453,7 +453,7 @@ se.gllvm <- function(object, ...){
               k<-max(y[,j])-2
               if(k>0){
                 sgns <- sign(zetas[(idx+1):(idx+k)])
-                cvs <- diag(sgns)%*%zeta.cov[(idx+1):(idx+k),(idx+1):(idx+k)]%*%diag(sgns)
+                cvs <- diag(sgns, nrow = k)%*%zeta.cov[(idx+1):(idx+k),(idx+1):(idx+k),drop=FALSE]%*%diag(sgns, nrow = k)
                 for(l in 1:k){
                   se.zetanew[j,l+1]<- sqrt(sum(cvs[l:k,l:k]))
                 } 
@@ -480,6 +480,7 @@ se.gllvm <- function(object, ...){
             names(se.zetanew)[1:2] <- c("cutoff0","cutoff1")
             se.zetanew <- se.zetanew[-((kz+ 1):length(se.zetanew))]
           }
+          sezetanew <- NULL
           if(any(family%in%c("ordinal"))){
             zetas <- object$TMBfn$par[names(object$TMBfn$par)=="zeta"]
             zeta.cov <- cov.mat.mod[(kz+ 1):(K-1),(kz+ 1):(K-1)]
@@ -490,7 +491,7 @@ se.gllvm <- function(object, ...){
               sezetanew <- c(sezetanew, sqrt(sum(cvs[1:i,1:i])))
             }
             if(any(object$family == "orderedBeta")){
-              se.zetanew <- c(se.zetanew[1:kz,1:kz],sezetanew)
+              se.zetanew <- c(se.zetanew[1:kz],sezetanew)
             }else{
               se.zetanew <- sezetanew
             }
@@ -601,14 +602,14 @@ se.gllvm <- function(object, ...){
     }
     
     if(method=="LA" || ((num.lv+num.lv.c)==0 && (method %in% c("VA", "EVA")) && is.null(object$params$row.params.random) && isFALSE(object$randomB)) && object$col.eff$col.eff!="random"){
-      covM <- try(MASS::ginv(sdr[incl,incl]))
-      if(inherits(covM, "try-error")) { stop("Standard errors for parameters could not be calculated, due to singular fit.\n") }
-      se <- try(sqrt(diag(abs(covM))))
+      cov.mat.mod <- try(MASS::ginv(sdr[incl,incl]))
+      if(inherits(cov.mat.mod, "try-error")) { stop("Standard errors for parameters could not be calculated, due to singular fit.\n") }
+      se <- try(sqrt(diag(abs(cov.mat.mod))))
       names(se) = names(object$TMBfn$par[incl])
       
       trpred<-try({
         if((num.lv+num.lv.c) > 0 || !is.null(object$params$row.params.random) || object$col.eff$col.eff == "random"){
-          sd.random <- sdrandom(objrFinal, covM, incl, ignore.u = FALSE)
+          sd.random <- sdrandom(objrFinal, cov.mat.mod, incl, ignore.u = FALSE)
           prediction.errors <- list()
           
           if(!is.null(object$params$row.params.random)){
@@ -636,7 +637,7 @@ se.gllvm <- function(object, ...){
       }, silent=TRUE)
       if(inherits(trpred, "try-error")) { cat("Prediction errors for random effects could not be calculated.\n") }
       
-      out$Hess <- list(Hess.full=sdr, incl=incl, cov.mat.mod=covM)
+      out$Hess <- list(Hess.full=sdr, incl=incl, cov.mat.mod=cov.mat.mod)
       
     } else {
       # cnrm <- apply(sdr,2,function(x)sqrt(sum(x^2)))
@@ -918,7 +919,7 @@ se.gllvm <- function(object, ...){
             if(attr(trm, "intercept"))LHS <- c("(Intercept)", LHS)
             RHS <- form[[3]]
             
-            names(sigma)[iter:(iter+trmsize[1,re]-1)] <- paste0(LHS, "|", RHS)
+            names(sigma)[iter:(iter+trmsize[1,re]-1)] <- paste0(LHS, "|", deparse(RHS))
             
             iter <- iter + trmsize[1,re]
           }
@@ -987,7 +988,7 @@ se.gllvm <- function(object, ...){
         se.zetanew <- matrix(NA,nrow=p,ncol=K)
         o_ind <- c(1:ncol(object$y))[family%in%c("ordinal", "orderedBeta")]
         zetas <- object$TMBfn$par[names(object$TMBfn$par)=="zeta"]
-        zeta.cov <- cov.mat.mod[names(object$TMBfn$par)[incl]=="zeta",names(object$TMBfn$par)[incl]=="zeta"]
+        zeta.cov <- cov.mat.mod[names(object$TMBfn$par)[incl]=="zeta",names(object$TMBfn$par)[incl]=="zeta", drop=FALSE]
         idx<-0
         for(j in o_ind){
           
@@ -995,7 +996,7 @@ se.gllvm <- function(object, ...){
             k<-max(y[,j])-2
             if(k>0){
               sgns <- sign(zetas[(idx+1):(idx+k)])
-              cvs <- diag(sgns)%*%zeta.cov[(idx+1):(idx+k),(idx+1):(idx+k)]%*%diag(sgns)
+              cvs <- diag(sgns, nrow = k)%*%zeta.cov[(idx+1):(idx+k),(idx+1):(idx+k),drop=FALSE]%*%diag(sgns, nrow = k)
               for(l in 1:k){
                 se.zetanew[j,l+1]<- sqrt(sum(cvs[l:k,l:k]))
               } 
@@ -1021,6 +1022,7 @@ se.gllvm <- function(object, ...){
           se.zetanew[2] <- object$params$zeta[2]*se.zetanew[2]
           names(se.zetanew)[1:2] <- c("cutoff0","cutoff1")
         }
+        sezetanew  <- NULL
         if(any(family%in%c("ordinal"))){
           zetas <- object$TMBfn$par[names(object$TMBfn$par)=="zeta"]
           zeta.cov <- cov.mat.mod[(kz+ 1):(K-1),(kz+ 1):(K-1)]
@@ -1032,7 +1034,7 @@ se.gllvm <- function(object, ...){
           }
         }
         if(any(object$family == "orderedBeta")){
-          se.zetanew <- c(se.zetanew[1:kz,1:kz],sezetanew)
+          se.zetanew <- c(se.zetanew[1:kz],sezetanew)
         }else{
           se.zetanew <- sezetanew
         }