+++ NEWS.md
@@ -3,13 +3,18 @@ Version 2.0.8
 
 * New predictSR function for predicting species richness
 * New poisson-binomial PMF function (for predictSR)
+* New residuals.predict.SR function for dunn-smyth residuals of species richness
 * New predictPairwise function (simple wrapper for predicting joint occurrence)
 * Implemented prediction functionality with when providing new measurements for covariates in the row effects
 * Improved output of AICc.gllvm to be in line with the AIC and BIC returned tables
+* AUC added to goodnessOfFit statistics
+* Added VA-EVA hybrid approximation for the orderedBeta model with logit link
 
 ## Bugfixes
 * row.eff was ignoring 'diag'
 * 2.0.7 did not install on ubuntu due to string error in new enum check
+* bugfix in plot.gllvm that left plots empty under particular settings
+* bugfix in nested random row effects
 
 
 Version 2.0.7